Slicer::Window: Fix a Gtk-CRITICAL situation
There's a bug in the Gtk::MenuButton interface.
set_popover takes a reference to a Popover.
The problem is that, when the MenuButton is destroyed, it tries to
destroy the Popover too.
Since it's a plain reference, not a Glib::RefPtr, the destruction fails
with a Gtk-CRITICAL.
It doesn't kill the program though.