Document: Fix two bugs related to rotated pages
The first one was realizing that PDFWriter sets page rotation as a page
attribute, so it was necessary to query the current rotation attribute,
and then adding to it the applied rotation.
The second one was that, when overwriting the current document with new
rotations, the following document saves would apply incorrect rotations
to the modified pages. This made necesarry to reload all Poppler related
stuff when overwriting the current document.